[libc++] Simplify the implementation of __tuple_leaf
