+++ libcxx/include/CMakeLists.txt
@@ -863,6 +863,7 @@ set(files
   __utility/as_lvalue.h
   __utility/auto_cast.h
   __utility/cmp.h
+  __utility/conditional_no_unique_address.h
   __utility/convert_to_integral.h
   __utility/declval.h
   __utility/empty.h
